added error handeling in component simulation

+++ src/Bess/include/scene/artist/nodes_artist.h
@@ -54,7 +54,7 @@ namespace Bess::Canvas {
                                  const Components::SpriteComponent &spriteComp,
                                  const Components::SimulationComponent &simComponent);
 
-        void paintSlot(uint64_t id, uint64_t parentId, const glm::vec3 &pos,
+        void paintSlot(int id, int parentId, const glm::vec3 &pos,
                        float angle, const std::string &label, float labelDx,
                        SimEngine::LogicState state, bool isConnected, SimEngine::ExtendedPinType extendedType) const;
